MailManager moved to interface/impl design.
