+++ app/src/main/resources/ApplicationResources.properties
@@ -312,7 +312,6 @@ configFrom.registration.disabled=Disabled -- No new accounts allowed
 # --------------------------------------------------------------- Comment emails
 
 email.comment.wrote=wrote
-email.comment.anonymous=An anonymous user wrote
 email.comment.reply=Reply to this comment at
 email.comment.management.link=Link to comment management page
 email.comment.commenter.email=Commenter's email address
