+++ app/src/main/resources/ApplicationResources_de.properties
@@ -197,7 +197,6 @@ weblogSettings.weblogUrl=URL
 editformat.html=Texteditor
 editor.rte.name=Rich Text Editor (Quill)
 # --------------------------------------------------------------- Entry editors
-email.comment.anonymous=Ein anonymer Benutzer schrieb
 email.comment.commentApproved=Kommentar zugelassen
 email.comment.reply=Beantworten Sie diesen Kommentar auf
 email.comment.title=Kommentar
