+++ app/src/main/resources/ApplicationResources_es.properties
@@ -132,7 +132,6 @@ weblogSettings.create.button.save=Crear weblog
 weblogSettings.error.invalidHandle=El t\u00EDtulo no es v\u00E1lido
 weblogSettings.error.handleExists=Ya existe un weblog con ese t\u00EDtulo
 email.comment.wrote=escribi\u00F3
-email.comment.anonymous=Un usuario an\u00F3nimo escribi\u00F3
 email.comment.reply=Respuesta a este comentario en
 email.comment.title=Comentario
 error.add.user.userNameInUse=El nombre de usuario ya est\u00E1 en uso
