+++ app/src/main/resources/ApplicationResources_fr.properties
@@ -244,7 +244,6 @@ editor.rte.name=Editeur de texte riche (Quill)
 # --------------------------------------------------------------- Comment emails
 
 email.comment.wrote=a \u00E9crit
-email.comment.anonymous=Un utilisateur anonyme a \u00E9crit
 email.comment.reply=R\u00E9pondre \u00E0 ce commentaire \u00E0
 email.comment.title=Commentaire
 email.comment.commentApproved=Commentaire approuv\u00E9
