+++ app/src/main/resources/ApplicationResources_ja.properties
@@ -202,7 +202,6 @@ weblogSettings.error.handleExists=\u305D\u306E\u30CF\u30F3\u30C9\u30EB\u540D\u30
 # --------------------------------------------------------------- Comment emails
 
 email.comment.wrote=\u3055\u3093\u304C\u4EE5\u4E0B\u306E\u30B3\u30E1\u30F3\u30C8\u3092\u6295\u7A3F\u3057\u307E\u3057\u305F
-email.comment.anonymous=\u533F\u540D\u30E6\u30FC\u30B6\u304C\u4EE5\u4E0B\u306E\u30B3\u30E1\u30F3\u30C8\u3092\u6295\u7A3F\u3057\u307E\u3057\u305F
 email.comment.reply=\u30B3\u30E1\u30F3\u30C8\u306B\u8FD4\u7B54\u3059\u308B
 email.comment.title=\u30B3\u30E1\u30F3\u30C8
 email.comment.commentApproved=\u30B3\u30E1\u30F3\u30C8\u304C\u627F\u8A8D\u3055\u308C\u307E\u3057\u305F
