+++ app/src/main/resources/ApplicationResources_ko.properties
@@ -240,7 +240,6 @@ editor.rte.name=\uB9AC\uCE58 \uD14D\uC2A4\uD2B8 \uD3B8\uC9D1\uAE30 (Quill)
 # --------------------------------------------------------------- Comment emails
 
 email.comment.wrote=\uC791\uC131
-email.comment.anonymous=\uC775\uBA85 \uC0AC\uC6A9\uC790 \uC791\uC131
 email.comment.reply=\uC774 \uC758\uACAC\uC5D0 \uB300\uD574 \uC751\uB2F5
 email.comment.title=\uC758\uACAC
 email.comment.commentApproved=\uC758\uACAC\uC774 \uC2B9\uC778\uB428
