+++ app/src/main/resources/ApplicationResources_ru.properties
@@ -160,7 +160,6 @@ weblogSettings.error.handleExists=\u0414\u043D\u0435\u0432\u043D\u0438\u043A \u0
 # --------------------------------------------------------------- Comment emails
 
 email.comment.wrote=\u043D\u0430\u043F\u0438\u0441\u0430\u0442\u044C
-email.comment.anonymous=\u0410\u043D\u043E\u043D\u0438\u043C\u043D\u044B\u0439 \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044C \u043D\u0430\u043F\u0438\u0441\u0430\u043B
 email.comment.reply=\u041E\u0442\u0432\u0435\u0442 \u043D\u0430 \u044D\u0442\u043E\u0442 \u043A\u043E\u043C\u043C\u0435\u043D\u0442\u0430\u0440\u0438\u0439
 email.comment.title=\u041A\u043E\u043C\u043C\u0435\u043D\u0442\u0430\u0440\u0438\u0439
