+++ app/src/main/resources/spring-beans.xml
@@ -145,9 +145,10 @@
         </beans:property>
     </beans:bean>
 
-    <beans:bean id="mailManager" class="org.apache.roller.weblogger.business.MailManager">
+    <beans:bean id="mailManager" class="org.apache.roller.weblogger.business.StringMailManager">
         <beans:constructor-arg ref="userManager"/>
         <beans:constructor-arg ref="weblogManager"/>
+        <beans:constructor-arg ref="weblogEntryManager"/>
         <beans:constructor-arg ref="propertiesManager"/>
         <beans:constructor-arg ref="urlStrategy"/>
         <beans:constructor-arg ref="mailSender"/>
