+++ app/src/main/resources/sql/createdb.vm
@@ -199,8 +199,8 @@ group by weblogid, name;
 create table weblog_entry_comment (
     id         varchar(48) not null primary key,
     entryid    varchar(48) not null,
-    name       varchar(255),
-    email      varchar(255),
+    name       varchar(255) not null,
+    email      varchar(255) not null,
     url        varchar(255),
     content    $db.TEXT_SQL_TYPE,
     posttime   $db.TIMESTAMP_SQL_TYPE not null,
